Align printer foreign key column types

+++ storage/migrations/2024_01_01_000016_create_printers.sql
@@ -3,7 +3,7 @@
 -- =====================================================
 
 CREATE TABLE IF NOT EXISTS printers (
-    id INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(150) NOT NULL,
     model VARCHAR(150),
     power_watts DECIMAL(10,2) DEFAULT 0,
